(fix) Fix block asc/desc ordering
It's still a bit weird becaus the blocks/list endpoint needs to be
kept in sync to the main dashboard rendering function.

But at least they are in sync now.

Maybe there is a way that does not require the sync but instead has
a single place where it is.